Update the blocks for crop and catch crop
