+++ fdm-app/app/routes/app.addfarm.$b_id_farm.cultivations.$b_lu_catalogue.covercrop.tsx
@@ -1,1 +1,10 @@
-export default function Index() {}
+export default function Index() {
+
+    return (<div className="space-y-6">
+        <div>
+            <p className="text-sm text-muted-foreground">
+                Vanggewas wordt binnenkort toegevoegd
+            </p>
+        </div>
+    </div>)
+}
